requirements.txt
trafilatura==1.12.2
readability-lxml==0.8.1
playwright==1.51.0
ebooklib==0.18
Pillow==11.1.0
lxml==5.3.1
click==8.1.8
ruff==0.9.10
# trafilatura transitive deps
babel==2.18.0
chardet==7.0.1
courlan==1.3.2
cssselect==1.4.0
dateparser==1.3.0
htmldate==1.9.4
justext==3.0.2
lxml-html-clean==0.4.4
python-dateutil==2.9.0.post0
pytz==2026.1.post1
regex==2026.2.28
six==1.17.0
tld==0.13.2
tzlocal==5.3.1
# playwright transitive deps
greenlet==3.3.2
pyee==12.1.1
typing-extensions==4.15.0
